+++ src/pages/contact/ContactForm.jsx
@@ -1,7 +1,6 @@
 import React from "react";
 import { Alert, Form } from "react-bootstrap";
 import { Formik } from "formik";
-import gql from "graphql-tag";
 import { Mutation } from "@apollo/react-components";
 import { Typeahead } from "react-bootstrap-typeahead";
 import { Query } from "react-apollo";
@@ -11,45 +10,13 @@ import { CONTAINER, FORM, BUTTON } from "../../components/StyledComponents";
 import Header from "../../components/header";
 import ContactFormFields from "../../components/contact/ContactFormFields";
 
+import {
+  GET_CONTACTS,
+  GET_CASES_TITLES
+} from "../../graphQL/queries"
+import { CREATE_CONTACT } from "../../graphQL/mutations"
+
 import { validContactSchema } from "../../validation";
-
-const CREATE_CONTACT = gql`
-  mutation CreateContact(
-    $firstName: String!
-    $lastName: String!
-    $caseRole: [Role]!
-    $email: String!
-    $courtCaseId: [ID]
-  ) {
-    createContact(
-      input: {
-        firstName: $firstName
-        lastName: $lastName
-        caseRole: $caseRole
-        email: $email
-        courtCaseId: $courtCaseId
-      }
-    ) {
-      contact {
-        id
-        firstName
-        lastName
-        caseRole
-        email
-      }
-      errors
-    }
-  }
-`;
-
-const GET_CASES_TITLES = gql`
-  query {
-    courtCases {
-      id
-      title
-    }
-  }
-`;
 
 class CreateContactForm extends React.Component {
   constructor(props) {
@@ -75,7 +42,16 @@ class CreateContactForm extends React.Component {
         ) : null}
 
         <CONTAINER>
-          <Mutation mutation={CREATE_CONTACT}>
+          <Mutation 
+            mutation={CREATE_CONTACT}
+            update={(cache, { data: { createContact } }) => {
+              // let c = cache.readQuery({ query: GET_CONTACTS });
+              // console.log(c)
+              // cache.writeQuery({
+              //   query: GET_CONTACTS,
+              //   data: { contacts: contacts.concat([createContact]) },
+              // });
+            }}>
             {(createContact, { data }) => (
               <div>
                 <Formik
@@ -105,7 +81,7 @@ class CreateContactForm extends React.Component {
                           alert: "Contact created successfully!",
                           type: "success"
                         });
-                        window.location.reload(false);
+                        //window.location.reload(false);
                       })
                       .catch(err => {
                         setSubmitting(false);
